Edited the Adelaide var
De-meaned the part rate variable before estimating in the var.  This seems to be a better specification as the deterministic term can be dropped from the var (as it was not siginificant) yet the part rate variable did have a non-zero mean (others didnt). VAR diagnostics look good and IRF looks good